requirements.txt
streamlit==1.38.0
google-generativeai==0.7.0
python-dotenv==1.0.1
langchain==0.3.0
PyPDF2==3.0.1
langchain_google_genai==2.0.0
langchain-community==0.3.0
faiss-cpu==1.8.0